requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg[binary]
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
python-multipart
email-validator
redis
